+++ changelog.md
@@ -4,6 +4,7 @@
 - Added ``getBlocksBetweenLocations(Location location1, Location location2)``, ``getSmallerLocation(Location location1, Location location2)``, ``getBiggerLocation(Location location1, Location location2)``, and ``isLocationUnderBlocks(Location location)``
 - Added Darkness effect
 - Added ``sendMessageToConsole(String pluginName, String message)``, ``sendWarningMessageToConsole(String pluginName, String message)`` and``sendErrorToConsole(String pluginName, String error, String explanation)``
+- Added ``createItem(Material type, int amount, String name, String loreString)`` which will enable adding multiple lines of lore using ``\n`` as a divider
 
 # Version 1.1